Add control-on-submit option and oncontrolchanged callback
On control elements, instead of firing events for both "change"
and "input" (which results in excessive events being fired),
if the control-on-submit option has been set,  fire events on
"change" (i.e. when the user is done editing) but otherwise
just fire on "input" (which is similar to existing behaviour
but without extra updates).

Additionally, after applying control changes, call oncontrolchanged.