minor changes to test-global
Fix plant_type to plantType to match convention.

Change the order of the plantType dropdown options so that
the initial setting (from the value in the data) isn't the
first setting, so as to be more sure it's getting set from
the data instead of just happening to be on the first option.