data-expands of string does single string
Change data-expands such that if the thing to be expanded is
a string, it's expanded once for the entire string and not
once for each character in the string.

BTW, strings are not instanceof String for javascriptish
reasons, but subclasses of strings would be.  I'm not
sure if this is a case we want to include.  For the moment,
I'm not including it.