INTERIM @key instead of data-shows-key
Kill data-shows-key;  instread, at expansion time, of @key is
specified either as data-shows or as an element value, substitute
the "key" (member name) into the expanded element.

This allows you to do fun and interesting things with <option>
tags as well as just showing the keys in relatively generic
cases of <div>s or whatever. (see test-options.html for fun cases).

The example code in the comments hasn't been tested, but otherwise
I think this is good.